feat: further apply cursor rules
Ensure that .types files are not in service layer
