+++ .gitignore
@@ -9,4 +9,4 @@ oclif.manifest.json
 yarn.lock
 pnpm-lock.yaml
 **/packages/cli-*/bin/
-**/tsconfig.tsbuildinfo\
+**/tsconfig.tsbuildinfo
