+++ src/api/types/nes.types.ts
