+++ src/commands/report/purls.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { Command, Flags, ux } from '@oclif/core';
 
-import type { Sbom } from '../../service/eol/eol.svc.ts';
+import type { Sbom } from '../../service/eol/cdx.svc.ts';
 import { extractPurls, getPurlOutput } from '../../service/purls.svc.ts';
 import SbomScan from '../scan/sbom.ts';
