--- src/service/nes/types.ts
@@ -1,5 +1,0 @@
-export type JSONValue = boolean | JSONValue[] | null | number | string | { [key: string]: JSONValue };
-
-export type Options = {
-  consent: boolean;
-};
