+++ test/commands/scan/eol.test.ts
@@ -5,10 +5,10 @@ import { afterEach, beforeEach, describe, it } from 'node:test';
 import { runCommand } from '@oclif/test';
 import * as sinon from 'sinon';
 
-import type { ScanResponseReport, ScanResult } from '../../../src/api/types/nes.ts';
+import type { ScanResponseReport, ScanResult } from '../../../src/api/types/nes.types.ts';
 import { default as SbomScan } from '../../../src/commands/scan/sbom.ts';
 import { type Sbom, cdxgen, prepareRows } from '../../../src/service/eol/eol.svc.ts';
-import type { CdxCreator } from '../../../src/service/eol/eol.types.ts';
+import type { CdxCreator } from '../../../src/service/eol/eol.svc.ts';
 import { buildScanResult } from '../../../src/service/nes/nes.svc.ts';
 import { extractPurls } from '../../../src/service/purls.svc.ts';
 import { FetchMock } from '../../utils/mocks/fetch.mock.ts';
