+++ test/service/line.svc.test.ts
@@ -1,6 +1,6 @@
 import assert from 'node:assert';
 import { describe, it } from 'node:test';
-import type { ScanResultComponent } from '../../src/api/types/nes.ts';
+import type { ScanResultComponent } from '../../src/api/types/nes.types.ts';
 import { daysBetween, formatLine, getMessageAndStatus, getStatusFromComponent } from '../../src/service/line.svc.ts';
 import type { Line } from '../../src/service/line.svc.ts';
