+++ test/service/nes.svc.test.ts
@@ -1,9 +1,9 @@
 import { ok } from 'node:assert';
 import { describe, it } from 'node:test';
 
-import type { ApolloHelper } from '../../../api/client.ts';
-import type { ScanResult, ScanResultComponent } from '../../../api/types/nes.ts';
-import { SbomScanner as sbomScanner } from '../nes.svc.ts';
+import type { ApolloHelper } from '../../src/api/client.ts';
+import type { ScanResult, ScanResultComponent } from '../../src/api/types/nes.types.ts';
+import { SbomScanner as sbomScanner } from '../../src/service/nes/nes.svc.ts';
 
 // TODO: move into /test/service/sbom.test.ts, but need to fix type errors first
 describe('SBOM Scanner', () => {
